# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Authentication API",
    "description": "API for user authentication",
    "version": "1.0.0"
  },
  "paths": {
...
    "/companies": {
      "get": {
        "tags": ["Companies"],
        "summary": "Get all companies",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of all companies",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Company"
              }
            }
          },
          "500": {
            "description": "Error fetching companies"
          }
        }
      },
      "post": {
        "tags": ["Companies"],
        "summary": "Create a new company",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "company",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewCompany"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Company created successfully",
            "schema": {
              "$ref": "#/definitions/Company"
            }
          },
          "400": {
            "description": "Missing required fields (name, address)"
          },
          "500": {
            "description": "Error creating company"
          }
        }
      },
      "options": {
        "tags": ["Companies"],
        "summary": "Get allowed methods for companies",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of allowed methods for the companies route"
          }
        }
      }
    },
    "/companies/{id}": {
      "get": {
        "tags": ["Companies"],
        "summary": "Get a single company by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the company to fetch"
          }
        ],
        "responses": {
          "200": {
            "description": "Company details",
            "schema": {
              "$ref": "#/definitions/Company"
            }
          },
          "404": {
            "description": "Company not found"
          },
          "500": {
            "description": "Error fetching company"
          }
        }
      },
      "put": {
        "tags": ["Companies"],
        "summary": "Update a company by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the company to update"
          },
          {
            "name": "company",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateCompany"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Company updated successfully"
          },
          "400": {
            "description": "Missing required fields (name, address)"
          },
          "500": {
            "description": "Error updating company"
          }
        }
      },
      "delete": {
        "tags": ["Companies"],
        "summary": "Delete a company by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the company to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Company deleted successfully"
          },
          "500": {
            "description": "Error deleting company"
          }
        }
      },
      "patch": {
        "tags": ["Companies"],
        "summary": "Partially update a company by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the company to update"
          },
          {
            "name": "updates",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateCompany"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Company partially updated successfully"
          },
          "400": {
            "description": "No fields provided to update"
          },
          "500": {
            "description": "Error partially updating company"
          }
        }
      },
      "options": {
        "tags": ["Companies"],
        "summary": "Get allowed methods for a specific company",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of allowed methods for the specific company route"
          }
        }
      }  
    },
    "/jobs": {
      "get": {
        "tags": ["Jobs"],
        "summary": "Get all jobs",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of all jobs",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Job"
              }
            }
          },
          "500": {
            "description": "Error fetching jobs"
          }
        }
      },
      "post": {
        "tags": ["Jobs"],
        "summary": "Create a new job",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "job",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewJob"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job created successfully",
            "schema": {
              "$ref": "#/definitions/Job"
            }
          },
          "400": {
            "description": "Missing required fields (title, description)"
          },
          "500": {
            "description": "Error creating job"
          }
        }
      }
    },
    "/jobs/{id}": {
      "get": {
        "tags": ["Jobs"],
        "summary": "Get a single job by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the job to fetch"
          }
        ],
        "responses": {
          "200": {
            "description": "Job details",
            "schema": {
              "$ref": "#/definitions/Job"
            }
          },
          "404": {
            "description": "Job not found"
          },
          "500": {
            "description": "Error fetching job"
          }
        }
      },
      "put": {
        "tags": ["Jobs"],
        "summary": "Update a job by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the job to update"
          },
          {
            "name": "job",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateJob"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job updated successfully"
          },
          "400": {
            "description": "Missing required fields (title, description)"
          },
          "500": {
            "description": "Error updating job"
          }
        }
      },
      "delete": {
        "tags": ["Jobs"],
        "summary": "Delete a job by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the job to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Job deleted successfully"
          },
          "500": {
            "description": "Error deleting job"
          }
        }
      },
      "patch": {
        "tags": ["Jobs"],
        "summary": "Partially update a job",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Updates specific fields of a job based on the provided job ID and request body.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the job to update",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "updates",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/JobUpdateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job updated successfully"
          },
          "400": {
            "description": "No fields provided for update"
          },
          "404": {
            "description": "Job not found"
          },
          "500": {
            "description": "Error updating the job"
          }
        }
      },
      "options": {
        "tags": ["Jobs"],
        "summary": "Provide allowed methods for the jobs endpoint",
        "description": "Returns the allowed HTTP methods for the jobs resource.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Allowed methods for the jobs endpoint",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "allowed_methods": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "example": {
                    "allowed_methods": ["GET", "POST", "OPTIONS"]
                  }
                }
              }
            }
          }
        }
      },
      "head": {
        "tags": ["Jobs"],
        "summary": "Check if a job exists",
        "description": "Checks if a job exists by its ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the job to check",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job exists"
          },
          "404": {
            "description": "Job not found"
          },
          "500": {
            "description": "Error checking the job"
          }
        }
      }
    },
...
    "/applications/job/{jobId}": {
      "get": {
        "tags": ["Applications"],
        "summary": "Fetch applications by job ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{
          "name": "jobId",
          "in": "path",
          "required": true,
          "type": "integer",
          "description": "Job ID"
        }],
        "responses": {
          "200": {
            "description": "List of applications for the job",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Application"
              }
            }
          },
          "500": {
            "description": "Error fetching applications for job"
          }
        }
      }
    },
...
    "/users/{id}/jobs": {
      "get": {
        "tags": ["Users"],
        "summary": "Get all jobs a user has applied for",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of jobs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/UserJob"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Company": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "sector": {
          "type": "string"
        },
        "contact_info": {
          "type": "object"
        },
        "address": {
          "type": "string"
        },
        "services": {
          "type": "object"
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "NewCompany": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "sector": {
          "type": "string"
        },
        "contact_info": {
          "type": "object"
        },
        "services": {
          "type": "object"
        }
      },
      "required": ["name", "address"]
    },
    "UpdateCompany": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "sector": {
          "type": "string"
        },
        "contact_info": {
          "type": "object"
        },
        "services": {
          "type": "object"
        }
      }
    },
    "Job": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "example": 1
        },
        "title": {
          "type": "string",
          "example": "Software Engineer"
        },
        "description": {
          "type": "string",
          "example": "Develop and maintain software applications."
        },
        "company_id": {
          "type": "integer",
          "example": 1
        }
      }
    },
    "NewJob": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "company_id": {
          "type": "integer"
        },
        "type": {
          "type": "string"
        }
      }
    },
    "UpdateJob": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "company_id": {
          "type": "integer"
        },
        "type": {
          "type": "string"
        } 
      }
    },
    "JobUpdateRequest": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        }
      }
    },
    "Application": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "user_id": {
          "type": "integer"
        },
        "job_id": {
          "type": "integer"
        },
        "application_date": {
          "type": "string",
          "format": "date-time"
        },
        "status": {
          "type": "string",
          "enum": ["pending", "accepted", "rejected"]
        }
      }
    },
...
    "UserJob": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter your Bearer token in the format: Bearer token",
      "x-display-name": "Bearer Token",
      "x-token-prefix": "Bearer "
    }
  }
}